requirements.txt
sqlalchemy
psycopg2-binary
pyTelegramBotApi
loguru
requests
BeautifulSoup4
newspaper3k